Chapter_3_Code/python_development/requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Werkzeug==2.0.3
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Flask==2.0.3
future==0.18.2
httpie==3.0.2
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
mypy-extensions==0.4.3
netmiko==3.4.0
nornir==3.2.0
nornir-netmiko==0.1.2
nornir-utils==0.1.2
ntc-templates==3.0.0
paramiko==2.9.2
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
PyYAML==6.0
pyyml==0.0.2
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scp==0.14.3
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
typing-extensions==3.10.0.2
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
